LPS-193651 Extract function to update the model builder structure
