+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/EditObjectFolder.tsx
@@ -3,13 +3,13 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {API, getLocalizableLabel} from '@liferay/object-js-components-web';
 import React, {useEffect, useState} from 'react';
 import {FlowElement} from 'react-flow-renderer';
 
 import {KeyValuePair} from '../ObjectDetails/EditObjectDetails';
 import {ModalAddObjectDefinition} from '../ViewObjectDefinitions/ModalAddObjectDefinition';
 import {ModalEditObjectFolder} from '../ViewObjectDefinitions/ModalEditObjectFolder';
+import {getUpdatedModelBuilderStructurePayload} from '../ViewObjectDefinitions/objectDefinitionUtil';
 import Diagram from './Diagram/Diagram';
 import EditObjectFolderHeader from './EditObjectFolderHeader/EditObjectFolderHeader';
 import {ModalPublishObjectDefinitions} from './EditObjectFolderHeader/ModalPublishObjectDefinitions';
@@ -62,115 +62,13 @@ export default function EditObjectFolder({
 			type: TYPES.SET_LOADING_OBJECT_FOLDER,
 		});
 
-		const makeFetch = async () => {
-			const objectFolders = await API.getAllObjectFolders();
-
-			const currentObjectFolder = objectFolders.find(
-				(objectFolder) => objectFolder.name === objectFolderName
-			) as ObjectFolder;
-
-			const objectFoldersWithObjectDefinitions: ObjectFolder[] = await Promise.all(
-				objectFolders.map(async (objectFolder) => {
-					const objectFolderWithObjectDefinitions: ObjectDefinitionNodeData[] = [];
-
-					const objectDefinitionsFilteredByObjectFolder = await API.getObjectDefinitions(
-						`filter=objectFolderExternalReferenceCode eq '${objectFolder.externalReferenceCode}'`
-					);
-
-					const linkedObjectDefinitions: ObjectDefinition[] = [];
-
-					await Promise.all(
-						objectFolder.objectFolderItems
-							.filter(
-								(objectFolderItem) =>
-									objectFolderItem.linkedObjectDefinition
-							)
-							.map(async (objectFolderItem) => {
-								linkedObjectDefinitions.push(
-									await API.getObjectDefinitionByExternalReferenceCode(
-										objectFolderItem.objectDefinitionExternalReferenceCode
-									)
-								);
-							})
-					);
-
-					const updateObjectFolderObjectDefinitions = ({
-						linkedObjectDefinition,
-						objectDefinitions,
-					}: {
-						linkedObjectDefinition: boolean;
-						objectDefinitions: ObjectDefinition[];
-					}) => {
-						objectDefinitions.forEach((objectDefinition) => {
-							const objectFolderItem = objectFolder.objectFolderItems.find(
-								(objectFolderItem) =>
-									objectFolderItem.objectDefinitionExternalReferenceCode ===
-									objectDefinition.externalReferenceCode
-							);
-
-							if (objectFolderItem) {
-								objectFolderWithObjectDefinitions.push({
-									...objectDefinition,
-									hasObjectDefinitionDeleteResourcePermission: !!objectDefinition
-										.actions.delete,
-									hasObjectDefinitionManagePermissionsResourcePermission: !!objectDefinition
-										.actions.permissions,
-									hasObjectDefinitionUpdateResourcePermission: !!objectDefinition
-										.actions.update,
-									hasObjectDefinitionViewResourcePermission: !!objectDefinition
-										.actions.get,
-									hasSelfObjectRelationships: false,
-									linkedObjectDefinition,
-									objectFields: objectDefinition.objectFields.map(
-										({
-											businessType,
-											externalReferenceCode,
-											label,
-											name,
-											required,
-										}) =>
-											({
-												businessType,
-												externalReferenceCode,
-												label: getLocalizableLabel(
-													objectDefinition.defaultLanguageId,
-													label,
-													name
-												),
-												name,
-												primaryKey: name === 'id',
-												required,
-												selected: false,
-											} as ObjectFieldNode)
-									),
-									selected: false,
-								});
-							}
-						});
-					};
-
-					updateObjectFolderObjectDefinitions({
-						linkedObjectDefinition: false,
-						objectDefinitions: objectDefinitionsFilteredByObjectFolder,
-					});
-
-					updateObjectFolderObjectDefinitions({
-						linkedObjectDefinition: true,
-						objectDefinitions: linkedObjectDefinitions,
-					});
-
-					return {
-						...objectFolder,
-						objectDefinitions: objectFolderWithObjectDefinitions,
-					};
-				})
+		const updateModelBuilderStructure = async () => {
+			const payload = await getUpdatedModelBuilderStructurePayload(
+				objectFolderName
 			);
 
 			dispatch({
-				payload: {
-					objectFolders: objectFoldersWithObjectDefinitions,
-					selectedObjectFolder: currentObjectFolder,
-				},
+				payload,
 				type: TYPES.UPDATE_MODEL_BUILDER_STRUCTURE,
 			});
 
@@ -182,7 +80,7 @@ export default function EditObjectFolder({
 			});
 		};
 
-		makeFetch();
+		updateModelBuilderStructure();
 
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [objectFolderName]);
