+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ViewObjectDefinitions/objectDefinitionUtil.d.ts
@@ -92,5 +92,11 @@ export declare function getObjectFolderActions(
 			type?: undefined;
 	  }
 )[];
+export declare function getUpdatedModelBuilderStructurePayload(
+	currentObjectFolderName: string
+): Promise<{
+	objectFolders: ObjectFolder[];
+	selectedObjectFolder: ObjectFolder;
+}>;
 export declare function normalizeName(str: string): string;
 export {};
